fftw: Fix miscompiled avx512 libraries
The intended compilation flags for AVX512 compilation are never set.
AVX512 libraries are always compiled with
   CFLAGS="$CFLAGS -march=haswell -mtune=haswell"
instead of the intended
   CFLAGS="$CFLAGS -march=skylake-avx512"

This is due to the fact that when we enable avx512 we always enable avx2 as well.
(i.e.: --enable-avx2,--enable-avx512).  However, we test for avx2 before we
test for avx512. Once we detect avx2, we don't test for avx512 anymore.
Hence we always assign avx2 flags for avx512 compilation.
In other words, we never hit the branch that sets CFLAGS as:
CFLAGS="$CFLAGS  -march=skylake-avx512"

The remedy is to change the order of tests: Test for avx512 before testing
for avx2.

Once in there, I also fixed somewhat confusing folder names: for avx2 and avx512
we build single float libraries in build folders named "build-double" and
double precission libraries are built in folders named "build-single".